% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round}
\alias{round}
\title{Round: Round numbers. Without outrage.}
\usage{
round(x, digits = 0)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{a positive integer indicating the number of decimal places. Negative integers will be converted to absolute values.}
}
\value{
a sensibly rounded number.
}
\description{
Round: Round numbers. Without outrage.
}
\examples{
x <- c(0.5, 1.5, 2.5, 3.5, 4.5)
round(x, digits = 0)
}
