% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round}
\alias{round}
\title{Round: Round numbers. Without outrage.}
\usage{
round(x, digits = 0)
}
\arguments{
\item{x}{a number. Just one. For now.}

\item{digits}{integer indicating the number of decimal places. Negative values are probably allowed, but I haven't tested this, so maybe just... don't.}
}
\value{
a sensibly rounded number.
}
\description{
Round: Round numbers. Without outrage.
}
\examples{
x <- 0.5
round(x, digits = 0)
}
